+++ src/main/java/entities/comment/CommonComment.java
@@ -3,7 +3,7 @@ package entities.comment;
 import entities.user.User;
 
 public class CommonComment implements Comment {
-    private final int ID;
+    private final long ID;
     private final String author;
     private final String content;
 
@@ -14,7 +14,7 @@ public class CommonComment implements Comment {
      * @param author  the author of the comment
      * @param content the content of the comment
      */
-    public CommonComment(int id, String author, String content) {
+    public CommonComment(long id, String author, String content) {
         ID = id;
         this.author = author;
         this.content = content;
@@ -27,7 +27,7 @@ public class CommonComment implements Comment {
      * @return its ID
      */
     @Override
-    public int getId() {
+    public long getId() {
         return ID;
     }
