+++ src/main/java/interface_adapter/schedule/ScheduleController.java
@@ -14,7 +14,7 @@ public class ScheduleController {
         this.scheduleInteractor = scheduleInteractor;
     }
 
-    public void getEvent(int projectId, int eventId, String name, String note, Date startAt, Date endAt, boolean isAllDay, List<String> userWith) {
+    public void getEvent(long projectId, long eventId, String name, String note, Date startAt, Date endAt, boolean isAllDay, List<String> userWith) {
         ScheduleInputData scheduleInputData = new ScheduleInputData(projectId, eventId, name, note, startAt, endAt, isAllDay, userWith);
         scheduleInteractor.getEvent(scheduleInputData);
     }
